HTTP user-agent uses the app_version now
Refactored some of the imports to avoid cyclical imports
