+++ core/exceptions.py
@@ -45,3 +45,31 @@ class BaseError(Exception):
         return "<BaseError(message={}, inner_exception={})>".format(
             (self), self.inner_exception
         )
+
+
+class IntegrationException(Exception):
+    """An exception that happens when the site's connection to a
+    third-party service is broken.
+
+    This may be because communication failed
+    (RemoteIntegrationException), or because local configuration is
+    missing or obviously wrong (CannotLoadConfiguration).
+    """
+
+    def __init__(self, message, debug_message=None):
+        """Constructor.
+
+        :param message: The normal message passed to any Exception
+        constructor.
+
+        :param debug_message: An extra human-readable explanation of the
+        problem, shown to admins but not to patrons. This may include
+        instructions on what bits of the integration configuration might need
+        to be changed.
+
+        For example, an API key might be wrong, or the API key might
+        be correct but the API provider might not have granted that
+        key enough permissions.
+        """
+        super().__init__(message)
+        self.debug_message = debug_message
